Refactoring state machine into separate modules
